Fixed jobs and added supermarket products list.

+++ StudentSimulator/Domain/Core/User/User.cs
@@ -10,7 +10,7 @@ namespace StudentSimulator.Domain.Core.User
         public double Weight {get; set;}
         public double Height {get; set;}
         public Characteristic Health {get;} = new(100, 100, 0);
-        public Characteristic Money {get;} = new(300, int.MaxValue, 0);
+        public Characteristic Money {get;} = new(300, double.MaxValue, 0);
         public Characteristic Mood {get;} = new(100, 100, 0);
         public Characteristic WaterLevel {get;} = new(2, 2, 0);
         public Characteristic EatLevel {get;} = new(2, 2, 0);
